Move connection resource to dedicated pkg
