+++ internal/auth0/connection/flatten_test.go
@@ -1,4 +1,4 @@
-package provider
+package connection
 
 import (
 	"testing"
