+++ internal/auth0/connection/resource_client_test.go
@@ -1,14 +1,12 @@
-package provider
+package connection_test
 
 import (
-	"context"
 	"fmt"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/stretchr/testify/assert"
 
+	"github.com/auth0/terraform-provider-auth0/internal/provider"
 	"github.com/auth0/terraform-provider-auth0/internal/recorder"
 	"github.com/auth0/terraform-provider-auth0/internal/template"
 )
@@ -68,7 +66,7 @@ func TestAccConnectionClient(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories: TestFactories(httpRecorder),
+		ProviderFactories: provider.TestFactories(httpRecorder),
 		Steps: []resource.TestStep{
 			{
 				Config: template.ParseTestName(testAccCreateConnectionClient, t.Name()),
@@ -102,55 +100,4 @@ func TestAccConnectionClient(t *testing.T) {
 			},
 		},
 	})
-}
-
-func TestImportConnectionClient(t *testing.T) {
-	var testCases = []struct {
-		testName             string
-		givenID              string
-		expectedConnectionID string
-		expectedClientID     string
-		expectedError        error
-	}{
-		{
-			testName:             "it correctly parses the resource ID",
-			givenID:              "conn_5678:client_1234",
-			expectedConnectionID: "conn_5678",
-			expectedClientID:     "client_1234",
-		},
-		{
-			testName:      "it fails when the given ID is empty",
-			givenID:       "",
-			expectedError: fmt.Errorf("ID cannot be empty"),
-		},
-		{
-			testName:      "it fails when the given ID does not have \":\" as a separator",
-			givenID:       "client_1234conn_5678",
-			expectedError: fmt.Errorf("ID must be formated as <connectionID>:<clientID>"),
-		},
-		{
-			testName:      "it fails when the given ID has too many separators",
-			givenID:       "client_1234:conn_5678:",
-			expectedError: fmt.Errorf("ID must be formated as <connectionID>:<clientID>"),
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.testName, func(t *testing.T) {
-			data := schema.TestResourceDataRaw(t, newConnectionClient().Schema, nil)
-			data.SetId(testCase.givenID)
-
-			actualData, err := importConnectionClient(context.Background(), data, nil)
-
-			if testCase.expectedError != nil {
-				assert.EqualError(t, err, testCase.expectedError.Error())
-				assert.Nil(t, actualData)
-				return
-			}
-
-			assert.Equal(t, actualData[0].Get("connection_id").(string), testCase.expectedConnectionID)
-			assert.Equal(t, actualData[0].Get("client_id").(string), testCase.expectedClientID)
-			assert.NotEqual(t, actualData[0].Id(), testCase.givenID)
-		})
-	}
 }
